# examples/out/swagger/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SWAGGER API",
    "contact": {
      "email": "hongdaesik88@gmail.com"
    },
    "license": {
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html",
      "name": "Apache 2.0"
    },
    "version": "0.0.1"
  },
  "servers": [
    {
      "url": "http://localhost",
      "description": "Dev"
    }
  ],
  "paths": {
    "/test/test": {
      "get": {
        "tags": [
          "test"
        ],
        "summary": "[100] function description",
        "responses": {
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "object",
                      "allOf": [
                        {
                          "$ref": "#components/schemas/Status"
                        }
                      ],
                      "title": "Status information *"
                    },
                    "param1": {
                      "type": "integer",
                      "format": "Int32",
                      "description": "int variable"
                    },
                    "param2": {
                      "type": "object",
                      "title": "data variable"
                    },
                    "param3": {
                      "type": "number",
                      "format": "Float",
                      "description": "float variable"
                    },
                    "param4": {
                      "type": "number",
                      "format": "Double",
                      "description": "double variable"
                    },
                    "param5": {
                      "type": "string",
                      "description": "string variable"
                    },
                    "param6": {
                      "type": "boolean",
                      "default": "false",
                      "description": "boolean variable"
                    },
                    "param7": {
                      "type": "object",
                      "allOf": [
                        {
                          "$ref": "#components/schemas/Response"
                        }
                      ],
                      "title": "struct variable"
                    },
                    "param8": {
                      "type": "integer",
                      "format": "Int32",
                      "description": "description of the value\n* 0= explain value\n* 1> explain value\n* 2< explain value\n* 3! explain value"
                    },
                    "param9": {
                      "type": "array",
                      "items": {
                        "type": "integer",
                        "format": "Int32"
                      },
                      "title": "int array variable"
                    },
                    "param10": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Parameter"
                      },
                      "title": "struct array variable"
                    }
                  }
                }
              }
            },
            "description": "응답 결과"
          }
        },
        "description": "### Relation.\n* [[101] function description](#test/post_test_test).\n### Description.\nparam1 mark variable explain",
        "parameters": [
          {
            "in": "query",
            "name": "param1",
            "schema": {
              "type": "integer",
              "format": "Int32",
              "description": "int variable",
              "nullable": true
            },
            "description": "int variable"
          },
          {
            "in": "query",
            "name": "param2",
            "schema": {
              "type": "object",
              "title": "data variable",
              "nullable": true
            },
            "description": "data variable"
          },
          {
            "in": "query",
            "name": "param3",
            "schema": {
              "type": "number",
              "format": "Float",
              "description": "float variable",
              "nullable": true
            },
            "description": "float variable"
          },
          {
            "in": "query",
            "name": "param4",
            "schema": {
              "type": "number",
              "format": "Double",
              "description": "double variable",
              "nullable": true
            },
            "description": "double variable"
          },
          {
            "in": "query",
            "name": "param5",
            "schema": {
              "type": "string",
              "description": "string variable",
              "nullable": true
            },
            "description": "string variable"
          },
          {
            "in": "query",
            "name": "param6",
            "schema": {
              "type": "boolean",
              "default": "false",
              "description": "boolean variable",
              "nullable": true
            },
            "description": "boolean variable"
          },
          {
            "in": "query",
            "name": "param7",
            "schema": {
              "type": "object",
              "allOf": [
                {
                  "$ref": "#components/schemas/Parameter"
                }
              ],
              "title": "struct variable",
              "nullable": true
            },
            "description": "struct variable"
          },
          {
            "in": "query",
            "name": "param8",
            "schema": {
              "type": "integer",
              "format": "Int32",
              "description": "description of the value\n* 0= explain value\n* 1> explain value\n* 2< explain value\n* 3! explain value",
              "nullable": true
            },
            "description": "description of the value\n* 0= explain value\n* 1> explain value\n* 2< explain value\n* 3! explain value"
          }
        ]
      },
      "post": {
        "tags": [
          "test"
        ],
        "summary": "[101] function description",
        "responses": {
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "object",
                      "allOf": [
                        {
                          "$ref": "#components/schemas/Status"
                        }
                      ],
                      "title": "Status information *"
                    },
                    "param1": {
                      "type": "integer",
                      "format": "Int32",
                      "description": "int variable"
                    },
                    "param2": {
                      "type": "object",
                      "title": "data variable"
                    },
                    "param3": {
                      "type": "number",
                      "format": "Float",
                      "description": "float variable"
                    },
                    "param4": {
                      "type": "number",
                      "format": "Double",
                      "description": "double variable"
                    },
                    "param5": {
                      "type": "string",
                      "description": "string variable"
                    },
                    "param6": {
                      "type": "boolean",
                      "default": "false",
                      "description": "boolean variable"
                    },
                    "param7": {
                      "type": "object",
                      "allOf": [
                        {
                          "$ref": "#components/schemas/Response"
                        }
                      ],
                      "title": "struct variable"
                    },
                    "param8": {
                      "type": "integer",
                      "format": "Int32",
                      "description": "description of the value\n* 0= explain value\n* 1> explain value\n* 2< explain value\n* 3! explain value"
                    },
                    "param9": {
                      "type": "array",
                      "items": {
                        "type": "integer",
                        "format": "Int32"
                      },
                      "title": "int array variable"
                    },
                    "param10": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Parameter"
                      },
                      "title": "struct array variable"
                    }
                  }
                }
              }
            },
            "description": "응답 결과"
          }
        },
        "description": "### Description.\nparam1 mark variable explain",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object",
                "required": [],
                "properties": {
                  "param1": {
                    "type": "integer",
                    "format": "Int32",
                    "description": "int variable"
                  },
                  "param2": {
                    "type": "object",
                    "description": "data variable"
                  },
                  "param3": {
                    "type": "number",
                    "format": "Float",
                    "description": "float variable"
                  },
                  "param4": {
                    "type": "number",
                    "format": "Double",
                    "description": "double variable"
                  },
                  "param5": {
                    "type": "string",
                    "description": "string variable"
                  },
                  "param6": {
                    "type": "boolean",
                    "default": "false",
                    "description": "boolean variable"
                  },
                  "param7": {
                    "type": "object",
                    "allOf": [
                      {
                        "$ref": "#components/schemas/Parameter"
                      }
                    ],
                    "description": "<details><summary>struct variable</summary><table><tbody><tr><td>param1</td><td><span>integer</span><span>($Int32)</span><div><p>int variable</p></div></td></tr><tr><td>param2</td><td><span>object</span><span></span><div><p>undefined</p></div></td></tr><tr><td>param3</td><td><span>number</span><span>($Float)</span><div><p>float variable</p></div></td></tr><tr><td>param4</td><td><span>number</span><span>($Double)</span><div><p>double variable</p></div></td></tr><tr><td>param5</td><td><span>string</span><span></span><div><p>string variable</p></div></td></tr><tr><td>param6</td><td><span>boolean</span><span></span><div><p>boolean variable</p></div></td></tr><tr><td>param7</td><td><span>integer</span><span>($Int32)</span><div><p>description of the value\n* 0= explain value\n* 1> explain value\n* 2< explain value\n* 3! explain value</p></div></td></tr></tbody></table></details>",
                    "example": {
                      "param1": null,
                      "param2": null,
                      "param3": null,
                      "param4": null,
                      "param5": null,
                      "param6": null,
                      "param7": null
                    }
                  },
                  "param8": {
                    "type": "integer",
                    "format": "Int32",
                    "description": "description of the value\n* 0= explain value\n* 1> explain value\n* 2< explain value\n* 3! explain value"
                  }
                }
              },
              "encoding": {
                "param2": {
                  "contentType": "application/json"
                },
                "param7": {
                  "contentType": "application/json"
                }
              }
            }
          }
        }
      }
    },
    "/test_path/test/{param1}": {
      "get": {
        "tags": [
          "test_path"
        ],
        "summary": "[200] function description",
        "responses": {
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "object",
                      "allOf": [
                        {
                          "$ref": "#components/schemas/Status"
                        }
                      ],
                      "title": "Status information *"
                    },
                    "param1": {
                      "type": "integer",
                      "format": "Int32",
                      "description": "int variable"
                    }
                  }
                }
              }
            },
            "description": "응답 결과"
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "param1",
            "schema": {
              "type": "integer",
              "format": "Int32",
              "description": "int variable *"
            },
            "required": true,
            "description": "int variable *"
          }
        ]
      }
    },
    "/test_path/test/{param1}/{param1}": {
      "put": {
        "tags": [
          "test_path"
        ],
        "summary": "[201] function description",
        "responses": {
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "object",
                      "allOf": [
                        {
                          "$ref": "#components/schemas/Status"
                        }
                      ],
                      "title": "Status information *"
                    },
                    "param1": {
                      "type": "integer",
                      "format": "Int32",
                      "description": "int variable"
                    }
                  }
                }
              }
            },
            "description": "응답 결과"
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "param1",
            "schema": {
              "type": "integer",
              "format": "Int32",
              "description": "int variable *"
            },
            "required": true,
            "description": "int variable *"
          }
        ],
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object",
                "required": [
                  "param2"
                ],
                "properties": {
                  "param2": {
                    "type": "string",
                    "description": "string variable *"
                  },
                  "param3": {
                    "type": "string",
                    "description": "string variable"
                  }
                }
              },
              "encoding": {}
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "token": []
    }
  ],
  "components": {
    "securitySchemes": {
      "token": {
        "in": "header",
        "type": "apiKey",
        "name": "X-API-TOKEN",
        "description": "\nDescription: 사용자 토큰값\n\n\n<details>\n<summary>Development token</summary>\n<p>Please enter the development token value.</p>\n</details>"
      }
    },
    "schemas": {
      "Status": {
        "required": [],
        "properties": {
          "code": {
            "type": "integer",
            "format": "Int32",
            "description": "Error code",
            "default": "null"
          },
          "message": {
            "type": "string",
            "description": "* 1: Korean description\n* 2: Korean description\n* 3: Korean description\n* 4: Korean description",
            "default": "null"
          }
        },
        "type": "object",
        "title": "Status ( Status information  )"
      },
      "Response": {
        "required": [],
        "properties": {
          "param1": {
            "type": "integer",
            "format": "Int32",
            "description": "int variable",
            "default": "null"
          },
          "param2": {
            "type": "object",
            "title": "data variable",
            "default": "null"
          },
          "param3": {
            "type": "number",
            "format": "Float",
            "description": "float variable",
            "default": "null"
          },
          "param4": {
            "type": "number",
            "format": "Double",
            "description": "double variable",
            "default": "null"
          },
          "param5": {
            "type": "string",
            "description": "string variable",
            "default": "null"
          },
          "param6": {
            "type": "boolean",
            "default": "null",
            "description": "boolean variable"
          },
          "param7": {
            "type": "object",
            "allOf": [
              {
                "$ref": "#components/schemas/Parameter"
              }
            ],
            "title": "struct variable",
            "default": "null"
          },
          "param8": {
            "type": "integer",
            "format": "Int32",
            "description": "description of the value\n* 0= explain value\n* 1> explain value\n* 2< explain value\n* 3! explain value",
            "default": "null"
          },
          "param9": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "Int32"
            },
            "title": "int array variable",
            "default": "null"
          },
          "param10": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Parameter"
            },
            "title": "struct array variable",
            "default": "null"
          }
        },
        "type": "object",
        "title": "Response ( Response information  )"
      },
      "Parameter": {
        "required": [],
        "properties": {
          "param1": {
            "type": "integer",
            "format": "Int32",
            "description": "int variable",
            "default": "null"
          },
          "param2": {
            "type": "object",
            "title": "data variable",
            "default": "null"
          },
          "param3": {
            "type": "number",
            "format": "Float",
            "description": "float variable",
            "default": "null"
          },
          "param4": {
            "type": "number",
            "format": "Double",
            "description": "double variable",
            "default": "null"
          },
          "param5": {
            "type": "string",
            "description": "string variable",
            "default": "null"
          },
          "param6": {
            "type": "boolean",
            "default": "null",
            "description": "boolean variable"
          },
          "param7": {
            "type": "integer",
            "format": "Int32",
            "description": "description of the value\n* 0= explain value\n* 1> explain value\n* 2< explain value\n* 3! explain value",
            "default": "null"
          }
        },
        "type": "object",
        "title": "Parameter ( Param information  )"
      }
    }
  }
}